fix: migrate check_mint_quote_paid fn from ln.rs to mod.rs
