+++ crates/cdk/src/mint/mod.rs
@@ -14,6 +14,8 @@ use cdk_common::nuts::{self, BlindSignature, BlindedMessage, CurrencyUnit, Id, K
 use cdk_common::payment::WaitPaymentResponse;
 pub use cdk_common::quote_id::QuoteId;
 use cdk_common::secret;
+use cdk_common::util::unix_time;
+use cdk_common::{MintQuoteState, PaymentMethod};
 use cdk_signatory::signatory::{Signatory, SignatoryKeySet};
 use futures::StreamExt;
 #[cfg(feature = "auth")]
@@ -37,7 +39,6 @@ mod builder;
 mod check_spendable;
 mod issue;
 mod keysets;
-mod ln;
 mod melt;
 mod proof_writer;
 mod start_up_check;
@@ -898,6 +899,71 @@ impl Mint {
         }
 
         Ok(total_redeemed)
+    }
+
+    /// Check the status of a payment for a quote
+    #[instrument(skip_all)]
+    pub async fn check_mint_quote_paid(&self, quote: &mut MintQuote) -> Result<(), Error> {
+        let state = quote.state();
+
+        // We can just return here and do not need to check with ln node.
+        // If quote is issued it is already in a final state,
+        // If it is paid ln node will only tell us what we already know
+        if quote.payment_method == PaymentMethod::Bolt11
+            && (state == MintQuoteState::Issued || state == MintQuoteState::Paid)
+        {
+            return Ok(());
+        }
+
+        let payment_processor = match self.payment_processors.get(&PaymentProcessorKey::new(
+            quote.unit.clone(),
+            quote.payment_method.clone(),
+        )) {
+            Some(pp) => pp,
+            None => {
+                tracing::info!("Could not get ln backend for {}, bolt11 ", quote.unit);
+
+                return Err(Error::UnsupportedUnit);
+            }
+        };
+
+        let payment_status = payment_processor
+            .check_incoming_payment_status(&quote.request_lookup_id)
+            .await?;
+
+        if payment_status.is_empty() {
+            return Ok(());
+        }
+
+        let mut tx = self.localstore.begin_transaction().await?;
+
+        for payment in payment_status {
+            if !quote.payment_ids().contains(&&payment.payment_id)
+                && payment.payment_amount > Amount::ZERO
+            {
+                tracing::debug!(
+                    "Found payment of {} {} for quote {} when checking.",
+                    payment.payment_amount,
+                    payment.unit,
+                    quote.id
+                );
+
+                let amount_paid = to_unit(payment.payment_amount, &payment.unit, &quote.unit)?;
+
+                quote.increment_amount_paid(amount_paid)?;
+                quote.add_payment(amount_paid, payment.payment_id.clone(), unix_time())?;
+
+                let total_paid = tx
+                    .increment_mint_quote_amount_paid(&quote.id, amount_paid, payment.payment_id)
+                    .await?;
+
+                self.pubsub_manager.mint_quote_payment(quote, total_paid);
+            }
+        }
+
+        tx.commit().await?;
+
+        Ok(())
     }
 }
